+++ CHANGELOG.md
@@ -21,6 +21,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [0.2.2] - 2026-03-02
+
+### Added
+
+- Add field visibility management and configuration
+- Add FieldSelector component and integrate with Header and NavMenu for managing visible fields
+
+### Changed
+
+- Improves accessibility in FieldSelector component by properly declaring as a modal and labelling it by the heading
+- Update ItemCard tests to use visibleFields and adjust field labels
+
+---
+
 ## [0.2.1] - 2026-03-02
 
 ### Changed